Added ordering to  getAllUsers endpoint.
To use the ordering you must send a request with the query string param orderColumn with the name the column which you want to order the table by. The orderClause is set by the query string param orderClause(if you send the orderColumn but omit the orderClause, by default, the value of the orderClause will be ASC).

For example:

localhost:8080/accounts/users/?orderColumn=rating&orderClause=DESC

localhost:8080/accounts/users/?orderColumn=name //orderClause=ASC