1-
Added the possibility to check if an account if following or being followed by an specific account.

To use this feature use the /account/{accountId}/followed endpoint (the same endpoint to obtain all accounts that are followed by this specific account), but instead of sending as query string the (pageSize and page or orderColumn or orderClause) you will send instead the query string pair accountToCheck={accountIdToCheck}.

The result will be a collection page with size 1 and with the followed account as item,if the relation exists, or will return an empty collection page if the relation do not exist.

The endpoint /account/{accountId}/following has the same logic of the previously explained endpoint, but instead of returning the accounts which the account with accountId is following will return the accounts that are following the account with accountId.

2- Fix hal links regarding the follow controller in output entities(OutAccount,OutUserAccount and OutCompany)