1- Fix follow tests.
2 -Changed the way the comments are created,get and updated.

GET /account/id/comments its supposed to return the comments MADE TO that account instead the comments made by that account.

To create a comment to an account  send a POST /account/{commentedAccountId}/comments , the account wich is making the comment is the account associated with the current authentication data.